refactor(api): move HMAC token + guard to src/common/hmac/
Both the existing benchmark callback path and the new RunCallbackController
(Phase 2 of #53) need the HMAC helpers. Shared common location avoids
duplication; the old benchmark/callbacks/ paths re-export from the new
location for compatibility through Phase 2.
