+++ apps/api/src/common/hmac/hmac-token.spec.ts
